Adding db generator

+++ OhmCalculatorApi/OhmCalculatorApi/Models/Color.cs
@@ -8,7 +8,7 @@ namespace OhmCalculatorApi.Models
         public int Id { get; set; }
         public string Rgb { get; set; }
         public string ValueDescription { get; set; }
-        public decimal ValueNumber { get; set; }
+        public double ValueNumber { get; set; }
         public ICollection<ColorSelectorConfiguration> ColorSelectorConfigurations { get; set; }
     }
 }
